requirements.txt
contextvars==2.4
Flask==1.1.2
pyquery==1.4.0
tqdm==4.47.0
numpy==1.18.5
Flask_Bootstrap==3.3.7.1
Flask_RESTful==0.3.8
aiohttp==3.6.2
redis==3.5.3
fake_useragent==0.1.11
opencv_python==4.3.0.36
requests==2.18.4
Pillow==7.2.0
